wip: Begin removing the async events channel
