+++ src/telemetry/metrics/serve.rs
@@ -1,21 +1,20 @@
-use deflate::CompressionOptions;
 use deflate::write::GzEncoder;
-use futures::future::{self, FutureResult};
-use http::{self, header, StatusCode};
-use hyper::{
-    service::Service,
-    Body,
-    Request,
-    Response,
+use deflate::CompressionOptions;
+use futures::{
+    future::{self, FutureResult}, Future,
 };
-
+use http::{self, header, StatusCode};
+use hyper::{service::Service, Body, Request, Response};
 use std::error::Error;
 use std::fmt;
 use std::io::{self, Write};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
+use tokio::executor::current_thread::TaskExecutor;
 
 use super::Root;
+use task;
+use transport::BoundPort;
 
 /// Serve Prometheues metrics.
 #[derive(Debug, Clone)]
@@ -42,12 +41,47 @@ impl Serve {
 
     fn is_gzip<B>(req: &Request<B>) -> bool {
         req.headers()
-            .get_all(header::ACCEPT_ENCODING).iter()
+            .get_all(header::ACCEPT_ENCODING)
+            .iter()
             .any(|value| {
-                value.to_str().ok()
+                value
+                    .to_str()
+                    .ok()
                     .map(|value| value.contains("gzip"))
                     .unwrap_or(false)
             })
+    }
+
+    pub fn serve(self, bound_port: BoundPort) -> impl Future<Item = (), Error = ()> {
+        use hyper;
+
+        let log = ::logging::admin().server("metrics", bound_port.local_addr());
+        let fut = {
+            let log = log.clone();
+            bound_port
+                .listen_and_fold(
+                    hyper::server::conn::Http::new(),
+                    move |hyper, (conn, remote)| {
+                        let service = self.clone();
+                        let serve = hyper.serve_connection(conn, service).map(|_| {}).map_err(
+                            |e| {
+                                error!("error serving prometheus metrics: {:?}", e);
+                            },
+                        );
+                        let serve = log.clone().with_remote(remote).future(serve);
+
+                        let r = TaskExecutor::current()
+                            .spawn_local(Box::new(serve))
+                            .map(move |()| hyper)
+                            .map_err(task::Error::into_io);
+
+                        future::result(r)
+                    },
+                )
+                .map_err(|err| error!("metrics listener error: {}", err))
+        };
+
+        log.future(fut)
     }
 }
 
@@ -66,8 +100,7 @@ impl Service for Serve {
             return future::ok(rsp);
         }
 
-        let mut metrics = self.metrics.lock()
-            .expect("metrics lock poisoned");
+        let mut metrics = self.metrics.lock().expect("metrics lock poisoned");
 
         metrics.retain_since(Instant::now() - self.idle_retain);
         let metrics = metrics;
@@ -85,7 +118,6 @@ impl Service for Serve {
                         .body(Body::from(body))
                         .map_err(ServeError::from)
                 })
-
         } else {
             let mut writer = Vec::<u8>::new();
             write!(&mut writer, "{}", *metrics)
@@ -125,7 +157,9 @@ impl From<io::Error> for ServeError {
 
 impl fmt::Display for ServeError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}: {}",
+        write!(
+            f,
+            "{}: {}",
             self.description(),
             self.cause().expect("ServeError must have cause")
         )
@@ -136,7 +170,7 @@ impl Error for ServeError {
     fn description(&self) -> &str {
         match *self {
             ServeError::Http(_) => "error constructing HTTP response",
-            ServeError::Io(_) => "error writing metrics"
+            ServeError::Io(_) => "error writing metrics",
         }
     }
